added rectangularChladniPattern and changed ints to longs

+++ src/physics/modes/linear_modes.hpp
@@ -17,7 +17,7 @@ namespace T = kac_core::types;
 
 namespace kac_core::physics {
 
-	inline T::Matrix_1D linearAmplitudes(const double& x, const int& N) {
+	inline T::Matrix_1D linearAmplitudes(const double& x, const unsigned long& N) {
 		/*
 		Calculate the amplitudes of the 1D eigenmodes relative to a strike
 		location.
@@ -30,11 +30,11 @@ namespace kac_core::physics {
 
 		T::Matrix_1D A(N);
 		double x_pi = x * pi;
-		for (unsigned int n = 0; n < N; n++) { A[n] = abs(sin((n + 1) * x_pi)); };
+		for (unsigned long n = 0; n < N; n++) { A[n] = abs(sin((n + 1) * x_pi)); };
 		return A;
 	}
 
-	inline T::Matrix_1D linearSeries(const int& N) {
+	inline T::Matrix_1D linearSeries(const unsigned long& N) {
 		/*
 		Calculate the the harmonic series.
 		input:
@@ -44,7 +44,7 @@ namespace kac_core::physics {
 		*/
 
 		T::Matrix_1D S(N);
-		for (unsigned int n = 0; n < N; n++) { S[n] = n + 1; };
+		for (unsigned long n = 0; n < N; n++) { S[n] = n + 1; };
 		return S;
 	}
